+++ src/plugins/curator/screens/gallery/index.jsx
@@ -1,3 +1,4 @@
+import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import AbstractNodeScreen from '@triniti/cms/plugins/ncr/screens/node';
 import createDelegateFactory from '@triniti/app/createDelegateFactory';
@@ -8,14 +9,21 @@ import selector from '@triniti/cms/plugins/curator/screens/gallery/selector';
 class GalleryScreen extends AbstractNodeScreen {
   static propTypes = {
     ...AbstractNodeScreen.propTypes,
+    formValues: PropTypes.object, // eslint-disable-line react/forbid-prop-types
   };
 
   static defaultProps = {
     ...AbstractNodeScreen.defaultProps,
+    formValues: {},
   };
 
   getForm() {
     return Form;
+  }
+
+  getFormRenderProps() {
+    const { formValues, getNodeRequestState } = this.props;
+    return { formValues, getNodeRequestState };
   }
 
   getTabs() {