+++ src/plugins/curator/screens/timeline/index.js
@@ -1,3 +1,4 @@
+import PropTypes from 'prop-types';
 import AbstractNodeScreen from '@triniti/cms/plugins/ncr/screens/node';
 import createDelegateFactory from '@triniti/app/createDelegateFactory';
 import { connect } from 'react-redux';
@@ -7,8 +8,23 @@ import selector from './selector';
 import schemas from './schemas';
 
 class TimelineScreen extends AbstractNodeScreen {
+  static propTypes = {
+    ...AbstractNodeScreen.propTypes,
+    formValues: PropTypes.object, // eslint-disable-line react/forbid-prop-types
+  };
+
+  static defaultProps = {
+    ...AbstractNodeScreen.defaultProps,
+    formValues: {},
+  };
+
   getForm() {
     return Form;
+  }
+
+  getFormRenderProps() {
+    const { formValues, getNodeRequestState } = this.props;
+    return { formValues, getNodeRequestState };
   }
 
   getTabs() {