+++ src/plugins/ovp/screens/video/index.jsx
@@ -1,3 +1,4 @@
+import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import AbstractNodeScreen from '@triniti/cms/plugins/ncr/screens/node';
 import createDelegateFactory from '@triniti/app/createDelegateFactory';
@@ -9,8 +10,23 @@ import selector from '@triniti/cms/plugins/ovp/screens/video/selector';
 import VideoAssetPreview from '@triniti/cms/plugins/dam/components/video-asset-preview';
 
 class VideoScreen extends AbstractNodeScreen {
+  static propTypes = {
+    ...AbstractNodeScreen.propTypes,
+    formValues: PropTypes.object, // eslint-disable-line react/forbid-prop-types
+  };
+
+  static defaultProps = {
+    ...AbstractNodeScreen.defaultProps,
+    formValues: {},
+  };
+
   getForm() {
     return Form;
+  }
+
+  getFormRenderProps() {
+    const { formValues, getNodeRequestState } = this.props;
+    return { formValues, getNodeRequestState };
   }
 
   getTabs() {