+++ src/plugins/ovp/screens/video/selector.js
@@ -1,3 +1,4 @@
+import { getFormValues } from 'redux-form';
 import getNode from '@triniti/cms/plugins/ncr/selectors/getNode';
 import updateScreenSelector from '@triniti/cms/plugins/ncr/screens/node/selector';
 import { formNames } from '../../constants';
@@ -12,5 +13,6 @@ export default (state, ownProps) => {
   return {
     ...updateScreenSelection,
     mezzanine: node && node.has('mezzanine_ref') ? getNode(state, node.get('mezzanine_ref')) : null,
+    formValues: getFormValues(formNames.VIDEO)(state),
   };
 };