+++ src/plugins/people/screens/person/Form.jsx
@@ -17,7 +17,7 @@ import setBlocks from '@triniti/cms/plugins/blocksmith/utils/setBlocks';
 import schemas from './schemas';
 
 const Form = ({
-  getNodeRequestState, node: person, form, tab, isEditMode, blocksmithState,
+  getNodeRequestState, node: person, form, formValues, tab, isEditMode, blocksmithState,
 }) => {
   switch (tab) {
     case 'details':
@@ -45,7 +45,7 @@ const Form = ({
     case 'taxonomy':
       return <TaxonomyFields schemas={schemas} isEditMode={isEditMode} />;
     case 'seo':
-      return schemas.node.hasMixin('triniti:common:mixin:seo') && <SeoFields isEditMode={isEditMode} />;
+      return schemas.node.hasMixin('triniti:common:mixin:seo') && <SeoFields formValues={formValues} isEditMode={isEditMode} />;
     case 'history':
       return (
         <History
@@ -94,6 +94,7 @@ Form.propTypes = {
     isDirty: PropTypes.bool.isRequired,
   }),
   form: PropTypes.string,
+  formValues: PropTypes.object, // eslint-disable-line react/forbid-prop-types
   getNodeRequestState: PropTypes.shape({
     request: PropTypes.instanceOf(Message).isRequired,
   }).isRequired,
@@ -105,6 +106,7 @@ Form.propTypes = {
 Form.defaultProps = {
   blocksmithState: null,
   form: null,
+  formValues: {},
   isEditMode: false,
   tab: '',
 };