+++ src/plugins/people/screens/person/index.jsx
@@ -17,11 +17,13 @@ class PersonScreen extends AbstractNodeScreen {
       editorState: PropTypes.instanceOf(EditorState),
       isDirty: PropTypes.bool.isRequired,
     }),
+    formValues: PropTypes.object, // eslint-disable-line react/forbid-prop-types
   };
 
   static defaultProps = {
     ...AbstractNodeScreen.defaultProps,
     blocksmithState: null,
+    formValues: {},
   };
 
   getBreadcrumbs() {
@@ -38,8 +40,8 @@ class PersonScreen extends AbstractNodeScreen {
   }
 
   getFormRenderProps() {
-    const { blocksmithState, getNodeRequestState } = this.props;
-    return { blocksmithState, getNodeRequestState };
+    const { blocksmithState, formValues, getNodeRequestState } = this.props;
+    return { blocksmithState, formValues, getNodeRequestState };
   }
 
   getForm() {