+++ src/plugins/taxonomy/screens/channel/selector.js
@@ -1,3 +1,4 @@
+import { getFormValues } from 'redux-form';
 import nodeScreenSelector from '@triniti/cms/plugins/ncr/screens/node/selector';
 import schemas from './schemas';
 import { formNames } from '../../constants';
@@ -5,4 +6,5 @@ import { formNames } from '../../constants';
 export default (state, ownProps) => nodeScreenSelector(state, ownProps, {
   schemas,
   formName: formNames.CHANNEL,
+  formValues: getFormValues(formNames.CHANNEL)(state),
 });