+++ CHANGELOG.md
@@ -1,11 +1,15 @@
 # CHANGELOG
 
 
+## vN.N.N
+* Adjust SEO Description Character Counter To Turn Yellow At 140 Characters and Red at 160.
+
+
 ## v0.13.4
 * Adding `alt_text` field to image asset forms.
 * Add eme-form-block to Blocksmith.
-* cms crash to white screen after paste
-* Use `@triniti/schemas` v1.1.15
+* CMS crash to white screen after paste.
+* Use `@triniti/schemas` v1.1.15.
 
 
 ## v0.13.3